fix(conversational_retrieval_chain.ts): keep the chat history context alongside the vector store context (#1801)
* fix(chains): update question_generator_template in chat_vector_db_chain.ts and conversational_retrieval_chain.ts

The `question_generator_template` in both `chat_vector_db_chain.ts` and `conversational_retrieval_chain.ts` files was updated to provide more clarity and include an optional chat history excerpt. The template now instructs to return the conversation history excerpt that includes any relevant context to the question if it exists and rephrase the follow up question to be a standalone question. The updated template also includes the phrase "Standalone question and optional chat history excerpt:".

* fix: remove changes from a deprecated chain

* chore: add ConversationalRetrievalQAChain tests and update its docs

* Remove images, update docs formatting

* Use context conversational retrieval QA prompt only in docs

* Format

* Update docs

---------

Co-authored-by: jacoblee93 <jacoblee93@gmail.com>