feat: add SerpAPILoader for web search result processing (#1972)
* feat: add SerpAPILoader and test cases

* feat(loader): improve URL building and response processing logic

* refactor(SerpAPILoader): Refactor buildUrl and processResponseData methods

Refactor the `buildUrl` and `processResponseData` methods in the `SerpAPILoader` class to improve code readability and maintainability.

* feat: Add example for using SerpAPI Loader in LangChain

This commit adds a new example file that demonstrates how to use the `SerpAPILoader` in LangChain to load web search results from SerpAPI. The example includes code that initializes the necessary components, defines a question and query, uses the loader to load the search results, stores the loaded documents in memory using `MemoryVectorStore`, and uses `RetrievalQAChain` to retrieve the most relevant documents and answer the question.

* feat(langchain): add serpapi document loader

* chore(examples/src/document_loaders): fix import in serpapi.ts

* fix: Update OpenAI and SerpAPI API keys

* fix(document_loaders): handle error message in SerpAPI loader

* fix: Update SerpAPILoader constructor to accept parameters object

The SerpAPILoader constructor is currently accepting apiKey and searchQuery parameters directly. This commit updates the constructor to accept a parameters object instead. The parameters object includes the q property for the search query. This change improves the readability and maintainability of the code.

* fix: fix indentation in serpapi.test.ts

* fix: Update SerpAPI Loader example

* Move apiKey into config param

* Remove from requires optional deps

* Fix entrypoints

---------

Co-authored-by: jacoblee93 <jacoblee93@gmail.com>