Automatically extract template variables  (#2334)
* chore(): add ExtractTemplateParamsRecursive and ParamsFrom types

* Update type in fromTemplate function

Updated the RunInput type in the fromTemplate method from 'never' to 'Symbol' within the 'prompt.ts' file. The reason for this change is to preserve the type safety in TypeScript when using Symbols as identifiers rather than 'never'. The corresponding return statement has also been adjusted in line with this change.

* Apply ppramesi's fix

* Update tests and docs

* Fix test

---------

Co-authored-by: jacoblee93 <jacoblee93@gmail.com>