Check a dev build of the docs, not a prod optimized build.
The `build` CI step was running a `yarn build` for everything, including docs. This is extremely slow (~15-17min) and unnecessary -- the unit tests already do a `yarn build` for the examples and for langchain itself, so the end result was duplication of work and an unnecessary (slow!) production build of the docs.

Here we rename the `build` CI step to `build-docs` and we test a dev build of the docs, which should be way faster.