feat: add momento vector index as a vector store provider
This adds Momento Vector Index (MVI) as a vector store provider. MVI
is part of the Momento platform. We have added Momento has a provider
for caching and session store previously. Because of this we update
the Momento package and peer dependencies.

We add integration tests and update existing tests for caching and
session store.