feat: add momento vector index as a vector store provider (#2851)
* feat: add momento vector index as a vector store provider

This adds Momento Vector Index (MVI) as a vector store provider. MVI
is part of the Momento platform. We have added Momento has a provider
for caching and session store previously. Because of this we update
the Momento package and peer dependencies.

We add integration tests and update existing tests for caching and
session store.

* chore: upgrade momento in examples and use `MOMENTO_API_KEY`

We upgrade the Momento SDK dependency in the examples and use the
environment variable `MOMENTO_API_KEY` which is the new standard in
Momento documentation.

* feat: update client type to interface for cross-envcompatibility

Modified the type from a specific client instance to an interface in
the cache, message store, and vector index. This change facilitates
dependency injection that's compatible across node, browser, and edge
environments.

* docs: document momento vector index and provide examples

* docs: document momento compatiblity in browser and edge environments

* docs: document vector store class

* refactor: renames momento-vector-index to momento_vector_index

Per style conventions, prefer snake case for modules over kebab
case. We update the module, tests, examples, docs, and name exports
per this convention.

We re-ran integration tests and examples to confirm working. We also
looked at the docs pages to confirm correct.