elasticsearch: customize user-agent for es client (#2897)
* elasticsearch: customize user-agent for es client

Updated the ElasticVectorSearch constructor to use a customized
user-agent for the elasticsearch client for telemetry.

* fix elasticsearch vectorstore int tests

Fixed the integration tests to work when not skipped, the before each
was lot setting the local store variable so it could be used in the
tests. I also removed any for the config to use the realy type so the
eslint disable was not needed.