Fixed parsing of tool_input and final_answer tags
When parsing the inputs and outputs from a tool, it is not properly renoving the end xml tag. For example for the input, the toolInput includes the </tool_input> which causes problems when passing this argument to the tool.

{
  tool: 'getLengthOfString',
  toolInput: 'This is a Happy String!</tool_input>',
  log: '<tool>getLengthOfString</tool><tool_input>This is a Happy String!</tool_input>'
}

Similarly, for the final_answer tag, it returns the end tags as well, for example:

{
  returnValues: {
    output: 'The length of the string "Parangaricutirimícuaro" is 22.</final_answer>'
  },
  log: '<final_answer>The length of the string "Parangaricutirimícuaro" is 22.</final_answer>'
}

I have change the code to use a regular expression and properly extract the text between the XML elements.