anthropic[minor]: fix nested object fixArrayXMLParameters
When parsing a nested object in fixArrayXMLParameters, the base schema
is passed on the recursive call. This causes a mismatch between the xml
object (the nested type) and the schema (still the base type)

I'm not very happy with type assertion for the schema, not sure what's
the correct check to make here though.