core[patch]: Fix "Bad control character in string literal" (#5160)
* Fixed "Bad control character in string literal"

The error "Bad control character in string literal" occurs when calling JSON.parse(json). This typically happens because the JSON string contains control characters that are not allowed, such as newline characters. Any control characters in the JSON string should be escaped before parsing it using JSON.parse().

* Added a test for PR#5135

Added a test for PR# #5135. Added a few newlines to the JSON text, which will result in an "Bad control character in string literal" error before the fix.

* Format

---------

Co-authored-by: jacoblee93 <jacoblee93@gmail.com>