feat(langchain): add orchestratorAbortBehavior option to RunnableConfig
This feature allows the consumer of a `Runnable` to define how the
orchestration layer (i.e. the LangChain core generic `Runnable`
implementations, like `RunnableLambda`, and LangGraph's task execution
logic) behaves when the abort signal is fired.

This is necessary because JavaScript doesn't provide any mechanism for
forcibly terminating an asynchronous task. The best you can do is hand
the task itself an `AbortSignal` and allow it to decide when to abort.
The only alternative to this would be to simply abandon the promise
instead of blocking until it settles.

Prior to this change, (and by default under this change) the execution
layer in LangChain will abandon a `Promise` and throw under most
circumstances when the `AbortSignal` is triggered. This means that
asynchronous work in progress that does not monitor the `AbortSignal`
will continue to execute, but the result of that work will not be
captured.

The change introduced by this feature allows the consumer of a runnable
to specify how eager the execution runtime should handle aborts. The
options are:

- `"throw_immediately"` (default): The execution runtime will abandon an
  active asynchronous task and throw when `RunnableConfig.signal`
  requests an abort.
- `"complete_pending"`: The execution runtime will *not* abandon
  active asynchronous tasks. Instead, the `AbortSignal` is checked
  before starting additional asynchronous tasks, and the execution
  runtime throws on abort only if it would otherwise start new
  asynchronous work.
- `"passthrough"`: The execution runtime will *not* check the
  `AbortSignal` at all in this case, but the signal will be passed
  through to the asynchronous tasks as per usual. This mode delegates
  the responsibility for early termination entirely to the asynchronous
  tasks themselves. This mode also makes it possible for execution
  runtimes external to LangChain core (like LangGraph) can request an
  abort while still allowing in-progress chains to complete so their
  state can be captured and replayed on resume.