fix: don't drop tool input when converting messages to Anthropic payload
As mentioned in https://github.com/langchain-ai/langchainjs/pull/7943#issuecomment-2807991300,
Anthropic's tool use inputs are Javascript objects at this point rather
than JSON strings, so we only need to parse the input if it's a string.

Without this change, the tool input is silently dropped when converting
messages due to the try/catch.