feat!(mcp-adapters): add authProvider for oauth, fix header passing
BREAKING_CHANGE: Drops the `useNodeEventSource` flag and stops passing
an empty eventSourceInit object to Streamable HTTP and SSE transport
constructors. If there continues to be header passing issues as a result
of the way `SSEClientTransport` and `StreamableHTTPClientTransport`
handle EventSource initialization, the correct fix will be to PR a
change to the MCP SDK that changes how headers are populated on
EventSource initialization.

fixes #8196
fixes #8194