fix(openai): bump zod 4 dep
This might seem like a misnomer for the zod 4 rollout, but test-exports-tsc is trying to shallow evaluate two different versions of zod (between langchain-core and langchain-openai) which quickly hits typescript recursion limits.

I've tested this independently and confirmed that zod deps disagreeing like this won't impact in a breaking way. `test-exports-tsc` fails because it resolves deps in a non-traditional way which causes the zod dep to resolve differently then how it does in practice.